upd(codegen): optionally generate README docs

+++ codegen.sh
@@ -5,7 +5,9 @@ _PWD=$PWD
 SPECS=$(find $PWD -name 'spec.json' | xargs -n1 dirname)
 for SPEC in $SPECS; do
 	echo "=== spec $SPEC ==="
-	cd $SPEC && rm -rf spec && /usr/bin/env spec && cd .. && ./README.php && cd $_PWD
+	if [ -x "$(dirname $SPEC)/README.php" ]; then
+		cd $SPEC && rm -rf spec && /usr/bin/env spec && cd .. && ./README.php && cd $_PWD
+	fi
 
 	SRC=$(dirname $(dirname $SPEC))
 	echo "=== codegen $SRC ==="
